Added Arcade Drive functionality
This has not been tested with the robot but it should work.
