+++ .pathplanner/settings.json
@@ -1,0 +1,12 @@
+{
+  "robotWidth": 1.0,
+  "robotLength": 1.0,
+  "holonomicMode": true,
+  "pathFolders": [],
+  "autoFolders": [],
+  "defaultMaxVel": 3.0,
+  "defaultMaxAccel": 3.0,
+  "defaultMaxAngVel": 540.0,
+  "defaultMaxAngAccel": 720.0,
+  "maxModuleSpeed": 1e+26
+}
