--- EverybotDrive/.wpilib/wpilib_preferences.json
@@ -1,6 +1,0 @@
-{
-    "enableCppIntellisense": false,
-    "currentLanguage": "java",
-    "projectYear": "2023",
-    "teamNumber": 6357
-}
