+++ simgui.json
@@ -1,0 +1,19 @@
+{
+  "HALProvider": {
+    "Other Devices": {
+      "SPARK MAX [33]": {
+        "header": {
+          "open": true
+        }
+      }
+    }
+  },
+  "NTProvider": {
+    "types": {
+      "/FMSInfo": "FMSInfo"
+    }
+  },
+  "NetworkTables Info": {
+    "visible": true
+  }
+}
